create separate modules, including route modules for features and shared
