+++ src/app/app.module.ts
@@ -1,29 +1,20 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { HttpClientModule } from '@angular/common/http';
-import { ExtraOptions, RouterModule, Routes } from '@angular/router';
 //
 import { AppComponent } from './app.component';
+import { AppRoutingModule } from './app-routing.module';
+//
 import { WelcomeComponent } from './home/welcome.component';
 import { ProductModule } from './products/product.module';
-// Warning: additional routes are defined within ProductModule
-const routes: Routes = [
-  { path: 'welcome', component: WelcomeComponent },
-  { path: '', redirectTo: 'welcome', pathMatch: 'full' },
-  { path: '**', redirectTo: 'welcome', pathMatch: 'full' },
-];
-
-const routeOptions: ExtraOptions = {
-  enableTracing: true,
-};
 
 @NgModule({
   declarations: [AppComponent, WelcomeComponent],
   imports: [
     BrowserModule,
     HttpClientModule,
-    RouterModule.forRoot(routes, routeOptions),
-    ProductModule,
+    ProductModule, // includes child routes
+    AppRoutingModule, // includes top-level routes, including wildcard route
   ],
 
   bootstrap: [AppComponent],
