+++ src/app/products/product-routing.module.ts
@@ -1,0 +1,23 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { Routes, RouterModule } from '@angular/router';
+//
+import { ProductDetailGuard } from './product-detail.guard';
+import { ProductDetailComponent } from './product-detail.component';
+import { ProductListComponent } from './product-list.component';
+
+const childRoutes: Routes = [
+  { path: 'products', component: ProductListComponent },
+  {
+    path: 'products/:id',
+    canActivate: [ProductDetailGuard],
+    component: ProductDetailComponent,
+  },
+];
+
+@NgModule({
+  declarations: [],
+  imports: [CommonModule, RouterModule.forChild(childRoutes)],
+  exports: [RouterModule],
+})
+export class ProductRoutingModule {}
