+++ src/app/products/product.module.ts
@@ -1,23 +1,13 @@
 // Product Feature Module
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { RouterModule, Routes } from '@angular/router';
 //
 import { SharedModule } from '../shared/shared.module';
 //
+import { ProductRoutingModule } from './product-routing.module';
 import { ProductListComponent } from './product-list.component';
 import { ProductDetailComponent } from './product-detail.component';
 import { ConvertToSpacesPipe } from '../shared/convert-to-spaces.pipe';
-import { ProductDetailGuard } from './product-detail.guard';
-
-const childRoutes: Routes = [
-  { path: 'products', component: ProductListComponent },
-  {
-    path: 'products/:id',
-    canActivate: [ProductDetailGuard],
-    component: ProductDetailComponent,
-  },
-];
 
 @NgModule({
   declarations: [
@@ -25,6 +15,6 @@ const childRoutes: Routes = [
     ProductDetailComponent,
     ConvertToSpacesPipe,
   ],
-  imports: [CommonModule, RouterModule.forChild(childRoutes), SharedModule],
+  imports: [CommonModule, SharedModule, ProductRoutingModule],
 })
 export class ProductModule {}
